Update documentation and enhance script options for ocp_report_pxesxi
This commit adds the `tenacity` dependency to the documentation and updates the default values for the `--credentials-secret` and `--credentials-namespace` options in the `ocp_report_pxesxi.py` script. Additionally, it introduces a new `--debug` option for enabling debug logging and clarifies the output description for ESXi hosts count. These changes aim to improve user guidance and enhance the functionality of the reporting script.