Enhance documentation for ocp_report_pxesxi and improve JSON output structure
This commit updates the `ocp_report_pxesxi.md` documentation to include a new `--vsphere-cert-path` option for specifying a custom SSL certificate file. It also clarifies the reporting of unique ESXi hosts in both the brief table and JSON outputs. Additionally, the Python script `ocp_report_pxesxi.py` is modified to improve the formatting of the brief JSON data generation function. These changes aim to enhance user guidance and improve the clarity of the reporting features.